move helpers types to separate file
