+++ common/api-review/firestore.api.md
@@ -12,7 +12,7 @@ import { LogLevelString as LogLevel } from '@firebase/logger';
 export function addDoc<T>(reference: CollectionReference<T>, data: WithFieldValue<T>): Promise<DocumentReference<T>>;
 
 // @public
-export type AddPrefixToKeys<Prefix extends string, T extends Record<string, any>> = {
+export type AddPrefixToKeys<Prefix extends string, T extends Record<string, unknown>> = {
     [K in keyof T & string as `${Prefix}.${K}`]+?: T[K];
 };
 
@@ -262,8 +262,8 @@ export { LogLevel }
 export function namedQuery(firestore: Firestore, name: string): Promise<Query | null>;
 
 // @public
-export type NestedUpdateFields<T extends Record<string, any>> = UnionToIntersection<{
-    [K in keyof T & string]: T[K] extends Record<string, any> ? AddPrefixToKeys<K, UpdateData<T[K]>> : never;
+export type NestedUpdateFields<T extends Record<string, unknown>> = UnionToIntersection<{
+    [K in keyof T & string]: T[K] extends Record<string, unknown> ? AddPrefixToKeys<K, UpdateData<T[K]>> : never;
 }[keyof T & string]>;
 
 // @public
@@ -471,7 +471,7 @@ export class Transaction {
 }
 
 // @public
-export type UnionToIntersection<U> = (U extends any ? (k: U) => void : never) extends (k: infer I) => void ? I : never;
+export type UnionToIntersection<U> = (U extends unknown ? (k: U) => void : never) extends (k: infer I) => void ? I : never;
 
 // @public
 export interface Unsubscribe {
